+++ src/TestUtils.js
@@ -1,7 +1,10 @@
-import AppCache from '../src/cache';
+import AppCache from './cache';
 
 //Used by tests
 function destroyAllDataPermanently() {
+  if (!process.env.TESTING) {
+    throw 'Only supported in test environment';
+  }
   return Promise.all(Object.keys(AppCache.cache).map(appId => {
       const app = AppCache.get(appId);
       if (app.databaseController) {
